Moved everything to main.c
