--- ADC_irq/Core/Inc/init.h
@@ -1,13 +1,0 @@
-/*
- * init.h
- *
- *  Created on: Apr 15, 2020
- *      Author: rodrigocelmer
- */
-
-#ifndef INC_INIT_H_
-#define INC_INIT_H_
-
-void ADC_Init(void);
-
-#endif /* INC_INIT_H_ */
